% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{clean_IGRA}
\alias{clean_IGRA}
\title{Turn NCDC raw igra data into R dataframe}
\usage{
clean_IGRA(path, type = "file")
}
\arguments{
\item{path}{character strings; Path to a file or folder.}

\item{type}{character strings; either 'file' or 'folder'.}
}
\value{
a dataframe if type == 'file'; a list of dataframes if type == 'folder'.
}
\description{
Turn NCDC raw igra data into R dataframe
}
\examples{
\dontrun{
data0 <- clean_IGRA('inst//extdata//v1//61902.y2d')
data1 <- clean_IGRA('inst//extdata//v1', type = 'folder')
}
}
