% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGRA_interface.R
\name{clean_IGRA_v1}
\alias{clean_IGRA_v1}
\title{Turn NCDC raw IGRA v1 data into an R dataframe}
\usage{
clean_IGRA_v1(file, type)
}
\arguments{
\item{file}{character strings; Path to a file. The file extension
should be one of ".y2d", ".dat", ".mly" and ".mly.201607".}

\item{type}{(optional) character; one of "y2d", "dat" and "mly". The
function will make a simple guess based on the filename if this is
not provided.}
}
\value{
a dataframe.
}
\description{
Turn NCDC raw IGRA v1 data into an R dataframe
}
\examples{
\dontrun{
path <- system.file("extdata//v1", package = "cleanIGRA")
data0 <- clean_IGRA_v1(file.path(path, '61902.y2d'))
data0 <- clean_IGRA_v1(file.path(path, '07139.dat'))
data0 <- clean_IGRA_v1(file.path(path, 'vwnd_12z.mly.201607'))
}
}
