% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGRA_interface.R
\name{clean_IGRA_v2}
\alias{clean_IGRA_v2}
\title{Turn NCDC raw IGRA v2 data into an R dataframe}
\usage{
clean_IGRA_v2(file, type)
}
\arguments{
\item{file}{character strings; Path to a file. The file extension
should be a ".txt" file.}

\item{type}{(optional) character; one of "data", "drvd" and "mly".
The function will make a simple guess based on the filename if this is
not provided.}
}
\value{
a dataframe.
}
\description{
Turn NCDC raw IGRA v2 data into an R dataframe
}
\examples{
\dontrun{
path <- system.file("extdata//v2", package = "cleanIGRA")
data0 <- clean_IGRA_v2(file.path(path, 'ASM00094703-data.txt'))
data0 <- clean_IGRA_v2(file.path(path, 'AGM00060490-drvd.txt'))
data0 <- clean_IGRA_v2(file.path(path, 'vapr_12z-mly-201906.txt'))
}
}
